requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
exceptiongroup==1.1.1
fonttools==4.38.0
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.0
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
plotly==5.13.1
plotly-express==0.4.1
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
statsmodels==0.13.5
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.0
tzdata==2022.7
urllib3==1.26.15
wordcloud==1.8.2.2
wsproto==1.2.0
zipp==3.15.0
